requirements.txt
flask
flask-cors
flask_sqlalchemy
requests
python-dotenv
pycryptodome
